Move library files around to groups that belong together (lib)
TODO: make it build on Windows and macOS
