+++ src/analytics.cc
@@ -14,10 +14,10 @@
 #include "const.h"
 #include "https_client.h"
 #include "platforminfo.h"
-#include "settings.h"
 #include "urls.h"
-#include "user.h"
 #include "util/logger.h"
+#include "model/settings.h"
+#include "model/user.h"
 
 namespace toggl {
