+++ src/batch_update_result.cc
@@ -5,7 +5,7 @@
 #include <sstream>
 #include <cstring>
 
-#include "base_model.h"
+#include "model/base_model.h"
 
 namespace toggl {
