+++ src/context.h
@@ -11,17 +11,17 @@
 #include <iostream> // NOLINT
 
 #include "analytics.h"
-#include "custom_error_handler.h"
 #include "feedback.h"
 #include "gui.h"
 #include "help_article.h"
 #include "idle.h"
-#include "util/logger.h"
 #include "model_change.h"
-#include "timeline_event.h"
 #include "timeline_notifications.h"
 #include "types.h"
 #include "websocket_client.h"
+#include "util/custom_error_handler.h"
+#include "util/logger.h"
+#include "model/timeline_event.h"
 
 #include <Poco/Activity.h>
 #include <Poco/LocalDateTime.h>
