+++ src/database/database.h
@@ -14,10 +14,10 @@
 
 #include <Poco/Data/SQLite/Connector.h>
 
-#include "model_change.h"
-#include "timeline_event.h"
-#include "types.h"
 #include "util/logger.h"
+#include "model/timeline_event.h"
+#include "model_change.h"
+#include "types.h"
 
 namespace Poco {
     namespace Data {
