+++ src/feedback.cc
@@ -4,8 +4,8 @@
 
 #include <sstream>
 
-#include "formatter.h"
 #include "https_client.h"
+#include "util/formatter.h"
 
 #include <Poco/Path.h>
