+++ src/gui.cc
@@ -6,16 +6,16 @@
 #include <cstdlib>
 #include <sstream>
 
-#include "client.h"
 #include "const.h"
 #include "error.h"
-#include "formatter.h"
-#include "project.h"
 #include "related_data.h"
-#include "task.h"
-#include "time_entry.h"
-#include "user.h"
-#include "workspace.h"
+#include "util/formatter.h"
+#include "model/client.h"
+#include "model/project.h"
+#include "model/task.h"
+#include "model/time_entry.h"
+#include "model/user.h"
+#include "model/workspace.h"
 
 #include <Poco/Stopwatch.h>
