+++ src/gui.h
@@ -10,11 +10,11 @@
 #include "help_article.h"
 #include "https_client.h"
 #include "proxy.h"
-#include "settings.h"
 #include "toggl_api.h"
 #include "toggl_api_private.h"
 #include "types.h"
 #include "util/logger.h"
+#include "model/settings.h"
 
 #include <Poco/LocalDateTime.h>
