+++ src/https_client.cc
@@ -7,10 +7,10 @@
 #include <string>
 #include <sstream>
 
-#include "formatter.h"
 #include "netconf.h"
 #include "urls.h"
 #include "toggl_api.h"
+#include "util/formatter.h"
 
 #include <Poco/DeflatingStream.h>
 #include <Poco/Environment.h>
