+++ src/model/autotracker.cc
