+++ src/model/base_model.cc
@@ -5,11 +5,11 @@
 #include <sstream>
 
 #include "batch_update_result.h"
-#include "database.h"
-#include "formatter.h"
 #include "model_change.h"
 #include "related_data.h"
 #include "user.h"
+#include "util/formatter.h"
+#include "database/database.h"
 
 #include <Poco/Timestamp.h>
 #include <Poco/DateTime.h>
