+++ src/model/settings.h
