+++ src/model/timeline_event.h
@@ -9,7 +9,7 @@
 #include <string>
 
 #include "base_model.h"
-#include "formatter.h"
+#include "util/formatter.h"
 
 namespace toggl {
