+++ src/model/workspace.cc
@@ -3,7 +3,7 @@
 #include "workspace.h"
 
 #include <sstream>
-#include "formatter.h"
+#include "util/formatter.h"
 
 namespace toggl {
