+++ src/related_data.cc
@@ -7,17 +7,17 @@
 
 #include <Poco/UTF8String.h>
 
-#include "autotracker.h"
-#include "formatter.h"
-#include "client.h"
 #include "gui.h"
-#include "obm_action.h"
-#include "project.h"
-#include "tag.h"
-#include "task.h"
-#include "time_entry.h"
-#include "user.h"
-#include "workspace.h"
+#include "util/formatter.h"
+#include "model/autotracker.h"
+#include "model/client.h"
+#include "model/obm_action.h"
+#include "model/project.h"
+#include "model/tag.h"
+#include "model/task.h"
+#include "model/time_entry.h"
+#include "model/user.h"
+#include "model/workspace.h"
 
 namespace toggl {
