+++ src/test/app_test.cc
@@ -4,22 +4,22 @@
 
 #include <iostream>  // NOLINT
 
-#include "./../autotracker.h"
-#include "./../client.h"
 #include "./../const.h"
-#include "./../database.h"
-#include "./../formatter.h"
-#include "./../obm_action.h"
-#include "./../project.h"
 #include "./../proxy.h"
-#include "./../settings.h"
-#include "./../tag.h"
-#include "./../task.h"
-#include "./../time_entry.h"
-#include "./../timeline_event.h"
 #include "./../timeline_uploader.h"
-#include "./../user.h"
-#include "./../workspace.h"
+#include "./../util/formatter.h"
+#include "./../database/database.h"
+#include "./../model/autotracker.h"
+#include "./../model/client.h"
+#include "./../model/obm_action.h"
+#include "./../model/project.h"
+#include "./../model/settings.h"
+#include "./../model/tag.h"
+#include "./../model/task.h"
+#include "./../model/time_entry.h"
+#include "./../model/timeline_event.h"
+#include "./../model/user.h"
+#include "./../model/workspace.h"
 
 #include "./test_data.h"
