+++ src/test/toggl_api_test.cc
@@ -6,12 +6,12 @@
 
 #include "toggl_api_test.h"
 #include "./../https_client.h"
-#include "./../obm_action.h"
 #include "./../proxy.h"
-#include "./../settings.h"
-#include "./../time_entry.h"
 #include "./../toggl_api.h"
 #include "./../toggl_api_private.h"
+#include "./../model/obm_action.h"
+#include "./../model/settings.h"
+#include "./../model/time_entry.h"
 #include "./test_data.h"
 
 #include <iostream>   // NOLINT
