+++ src/toggl_api_private.h
@@ -6,12 +6,12 @@
 #include <string>
 #include <vector>
 
-#include "autotracker.h"
 #include "help_article.h"
 #include "proxy.h"
-#include "settings.h"
 #include "toggl_api.h"
 #include "util/logger.h"
+#include "model/autotracker.h"
+#include "model/settings.h"
 
 namespace toggl {
 class Client;
