+++ src/util/custom_error_handler.h
