+++ src/util/rectangle.cc
