+++ src/util/rectangle.h
